# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeLedger API specification for Australian sole trader tradies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads/receipt:
    post:
      operationId: uploadReceipt
      tags: [uploads]
      summary: Upload a receipt image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                receipt:
                  type: string
                  format: binary
              required:
                - receipt
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadReceiptResponse"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    UploadReceiptResponse:
      type: object
      properties:
        fileUrl:
          type: string
        filename:
          type: string
          nullable: true
      required:
        - fileUrl
...